requirements.txt
matplotlib~=3.0.3
numpy
pillow~=5.4.1